Create data_parsing.cpp
Started writing out code for data parsing of airport data set.  Very much high level. Need file specifics and location. And need to figure out exact graph variables to store info. Also made this a void function for now, may be better to put it as part of main but that is a conversation for latter.